+++ SDK/AppCenterDistribute/Microsoft.AppCenter.Distribute.iOS.Bindings/ApiDefinition.cs
@@ -44,20 +44,30 @@ namespace Microsoft.AppCenter.Distribute.iOS.Bindings
         [Export("notifyUpdateAction:")]
         void NotifyUpdateAction(MSUpdateAction action);
 
-        // + (void)setUpdateTrack:(MSUpdateTrack)updateTrack
+        // + (void)setUpdateTrack:(MSUpdateTrack)updateTrack;
         [Static]
         [Export("setUpdateTrack:")]
         void SetUpdateTrack(MSUpdateTrack updateTrack);
 
-        // + (MSUpdateTrack)updateTrack
+        // + (MSUpdateTrack)updateTrack;
         [Static]
         [Export("updateTrack")]
         MSUpdateTrack GetUpdateTrack();
 
-        // +(void)resetSharedInstance
+        // +(void)resetSharedInstance;
         [Static]
         [Export("resetSharedInstance")]
         void ResetSharedInstance();
+
+        // + (void)checkForUpdate;
+        [Static]
+        [Export("checkForUpdate")]
+        void CheckForUpdate();
+
+        // + (void)disableAutomaticCheckForUpdate;
+        [Static]
+        [Export("disableAutomaticCheckForUpdate")]
+        void DisableAutomaticCheckForUpdate();
     }
 
     // @protocol MSDistributeDelegate <NSObject>